feat(ev): 속도별 요금 표시 + 요금 소수점 보존
서버가 출력 구간별 회원요금(tierPrices)을 내려준다. 같은 충전소라도 50kW와
100kW 요금이 다른데(채비 410/430) 지금까지 하나로만 보여줬다. 요금 탭에
'속도별 회원 요금' 카드를 추가 — 그 충전소에 실제 있는 구간만, 2개 이상일 때만.

요금을 int 로 받으면서 .round() 로 뭉개고 있었다. 고시가가 307.2·325.6·348.4·
393.1 처럼 소수 한 자리라 348.4가 348로 보였다. 모델을 double 로 바꾸고
표시부(카드·상세·AI 결과 화면)를 evPriceText 로 통일했다. 총 예상금액은
기존대로 정수+콤마 — 단가만 소수를 유지한다.

작은 화면 방어: 소수점이 붙어 '393.1원'으로 두 자 길어졌다. 요금 카드 숫자와
속도별 금액을 FittedBox(scaleDown)로 감싸 칸 안에서 줄어들게 하고, 라벨은
Expanded + ellipsis. '길 안내 시작 (12.3Km)' 버튼도 한 줄 + 축소 — 지도 시트가
목록 거리를 이어받으면서 이 라벨이 시트에도 나오게 됐다.

helpers.dart 의 formatEvPrice 는 제거했다. 사용처가 0인데 .toInt() 로 잘라서
소수점 요구와 정면으로 충돌하는 함정이었다.

+++ lib/ui/favorites/favorites_screen.dart
@@ -55,7 +55,7 @@ class _FavoritesScreenState extends ConsumerState<FavoritesScreen>
 
   List<EvStation> _sortedEv(List<EvStation> l) {
     final c = [...l];
-    int? priceOf(EvStation s) =>
+    double? priceOf(EvStation s) =>
         s.unitPriceFast ??
         s.unitPriceSlow ??
         s.unitPriceFastMember ??
